moves composite components into groups folder
