+++ src/groups/SchematicDiagram/SchematicDiagram.scss
