+++ src/groups/SchematicDiagram/SchematicDiagram.tsx
