+++ src/groups/SchematicDiagram/colors.ts
