+++ src/groups/SchematicDiagram/diagramLayer.ts
