+++ src/groups/SchematicDiagram/index.ts
